# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProyectoTech",
        "description": "Documentación proyectoTech (CRUD de productos y carga masiva - Carrito de Compras)",
        "contact": {
            "email": "johan33@hotmail.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Desarrollo ProyectoTech"
        }
    ],
    "paths": {
...
        "/auth/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "description": "Servicio para visualizar todos los productos registrados en la base de datos",
                "operationId": "productos",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/productos/nuevo": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Nuevo producto",
                "description": "Servicio para el registro de un nuevo producto",
                "operationId": "productos/nuevo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "unidades"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del producto a crear",
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Nombre del producto"
                                    },
                                    "precio": {
                                        "description": "Precio del producto",
                                        "type": "float",
                                        "example": 150000
                                    },
                                    "unidades": {
                                        "description": "Unidades disponibles del producto",
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/productos/buscar/{producto}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Consultar producto por ID",
                "description": "Servicio para consultar un producto en específicdo",
                "operationId": "productos/buscar/{producto}",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/productos/actualizar/{producto}": {
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "description": "Servicio para la actualización de datos de un producto",
                "operationId": "productos/actualizar/{producto}",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del producto modificado",
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Nombre del producto"
                                    },
                                    "precio": {
                                        "description": "Precio del producto modificado",
                                        "type": "float",
                                        "example": 150000
                                    },
                                    "unidades": {
                                        "description": "Unidades disponibles del producto modificadas",
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/productos/eliminar/{producto}": {
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "description": "Servicio para eliminar un producto",
                "operationId": "productos/eliminar/{producto}",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/productos/carga": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Carga masiva productos",
                "description": "Servicio para el registro masivo de productos por medio de un archivo CSV",
                "operationId": "productos/carga",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "archivo": {
                                        "title": "archivo",
                                        "description": "Archivo CSV",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transacción exitosa",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "400": {
                        "description": "Petición no válida"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "500": {
                        "description": "Error de servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}